requirements.txt
pandas
pillow 
openpyxl
email
asyncio
pyopenssl